Save zeta as a int16
Bench: 11382784

+++ Halogen/src/Position.h
@@ -75,7 +75,7 @@ private:
 	deltaArray delta;										//re recycle this object to save time in CalculateMoveDelta
 
 	//Values for keeping the network updated
-	std::array<std::array<int32_t, HIDDEN_NEURONS>, MAX_DEPTH> Zeta;
+	std::array<std::array<int16_t, HIDDEN_NEURONS>, MAX_DEPTH> Zeta;
 	size_t incrementalDepth = 0;
 };
